requirements.txt
numpy
pandas
pymupdf
pillow
ipython
openpyxl
matplotlib
seaborn
opencv-python
torch
torchvision
tqdm
requests
scikit-learn
scikit-image